+++ composition-js/src/__tests__/compose.test.ts
@@ -1227,6 +1227,33 @@ describe('composition', () => {
       ]);
     });
 
+    it('errors when a subgraph has a field with an introspection-reserved name', () => {
+      const subgraphA = {
+        typeDefs: gql`
+          type Query {
+            __someQuery: Int
+          }
+        `,
+        name: 'subgraphA',
+      };
+
+      const subgraphB = {
+        typeDefs: gql`
+          type Query {
+            aValidOne: Int
+          }
+        `,
+        name: 'subgraphB',
+      };
+
+      const result = composeAsFed2Subgraphs([subgraphA, subgraphB]);
+
+      expect(result.errors).toBeDefined();
+      expect(errors(result)).toStrictEqual([
+        ['INVALID_GRAPHQL', '[subgraphA] Name "__someQuery" must not begin with "__", which is reserved by GraphQL introspection.'],
+      ]);
+    });
+
     it('errors when the @tag definition is invalid', () => {
       const subgraphA = {
         typeDefs: gql`
@@ -2751,6 +2778,44 @@ describe('composition', () => {
       expect(printType(result.schema.toAPISchema().type('E')!)).toMatchString(`
         enum E {
           V1
+        }
+      `);
+    });
+
+    it('do not error if a skipped inconsistent value has directive applied', () => {
+      const subgraphA = {
+        name: 'subgraphA',
+        typeDefs: gql`
+          type Query {
+            f(e: E!): Int
+          }
+
+          enum E {
+            V1
+            V2 @deprecated(reason: "use V3 instead")
+            V3
+          }
+        `,
+      };
+
+      const subgraphB = {
+        name: 'subgraphB',
+        typeDefs: gql`
+          enum E {
+            V1
+            V3
+          }
+        `,
+      };
+
+      const result = composeAsFed2Subgraphs([subgraphA, subgraphB]);
+      assertCompositionSuccess(result);
+      // Note that we will also generate an hint for the skipped value but this is already
+      // tested by `hints.test.ts` so we don't duplicate the check here.
+      expect(printType(result.schema.toAPISchema().type('E')!)).toMatchString(`
+        enum E {
+          V1
+          V3
         }
       `);
     });